Add SDL input handling
Input is now handled by a SDL_PollEvent loop and is much more reliable
than before. Opcode 0xFx0A still needs to be fixed to only proceed when
a key is released. The key registers are cleared after each time they are
consumed by an opcode, but at no other time. It may make sense to also
clear them on the same interval as the draw timer (60 Hz); more testing
needs to be done.