Add menubar, file opening dialog, reset
WinAPI is used for the GUI elements for the sake of being lightweight
and easy to use with SDL. The implementation of the menus sits in
the emuWindow class - with event handling in the SDL_PollEvent loop -
but these may need to be spun out in the future. The const enum buttonIDs
are in the emuWindow class scope (in the definition file) and in the
global scope of chip8emu - there is a better solution for this that needs
to be implemented in the future. These are used in event handling and
for appending to the menus.