Add screen color changing
Implemented the ability to pick colors for the "on" and "off" pixels.
There are a few presets based on historical monochrome displays, but
there is also the ability to pick any colors using the WinAPI color
picker. This was all added as a part of the emuWindow class to keep it
separate from the Chip-8 functionality and closer to the GUI code.