Add sound system
Sound is played through the SDL_audio library using the beeper helper
class. After checking the sound timer in the chip-8, the main emulation
loop will either give a true or false value to the beeper class in the
setSoundOn(bool) which will have the beeper feed wave values or silence
into the SDL audio buffer. Currently, users can switch between a sine or
a square wave. In the future, users will also be able to set the frequency
of the sound.