Port GUI, event handling to wxWidgets
wxWidgets is now used to handle the GUI and events, taking over from SDL.
SDL is still used for drawing pixels on the screen and for sound.

wxWidgets takes over from main(), so there is no longer a visible core
loop. Rendering is done using the onIdle() event callback function in
cMain, which contains a timer. The Chip-8 CPU cycling loop is spawned in
a different thread when it gets started, with new options to only cycle
once or to loop. Doing this will allow for a frame-by-frame mode in the
future.

The cApp class (extends wxApp, wraps everything) contains the cMain class,
which contains the chip-8 instance, sound maker, and the SDL window and
renderer.