Add state save/load
State saving and loaded are now implemented. State saving is done by
"getting" all the necessary pieces of the Chip-8 from the class - that is,
the memory, registers, "I" register, program counter, stack, stack
pointer, delay and sound timers, and screen buffer - and writing that as
binary to the user-chosen state file (prefix .c8s). This is all handled
by the cMain class.

State loading is done by first reading the entire user-chosen state file
into an array of unsigned chars. This part is handled by the cMain class.
This array is then handed off to the Chip-8 class, which iterates through
the array and puts the relevant data into the correct variable based on
the position in the array.

The memory, registers, and screen are simple, since they are saved and
loaded as unsigned chars. The "short" typed variables are more complicated
and on of their bytes must be bit-shifted. NOTE: The current
implementation was made on a Windows machine which is little-endian. This
may cause problems on other machines and an endian-neutral implementation
should be provided later.