Add config.txt and keybinding dialog
On start-up, the emulator looks for "config.txt" in the local directory
and reads in the keybindings. Right now, only keybindings are allowed in
the config file and it does not update when keybindings are changed while
the program is open. In the future, all configuration options will be
in the file (graphics, sound, rom/savestate directories) and it will be
updated when changes are made in the program at runtime.

The keybinding dialog is passed a reference to the map which associates
Chip-8 keys (as ints from 0 to 15) to Unicode decimal values. Similar to
other simple keybinding dialogs in games, clicking on a Chip-8 control
will "unbind" it and the dialog will wait for a key press event to
"rebind" the control. Rebinding to a key already used for a control will
swap the bindings and ok-ing the dialog before rebinding a control will
leave that control unchanged.