Add memory viewer, emulation speed controls
The memory viewer is implemented in a child window using custom
virtual wxListCtrl components. These components store pointers to the
data they are responsible for showing, which means the main window
doesn't have to be responsible for pushing new data to the memory viewer
on each render loop. However, the call for the ListCtrls to update
happens each render loop, so usually multiple Chip-8 cycles have passed
since the last update on the components.

Emulation speed could always be changed inside the Chip-8 class, but now
a menu item and event handling exists for that purpose in the main
window, along with the ability to allow one Chip-8 cycle to pass and stop.
This is really useful for debugging and makes the memory viewer usable.

Along with the above changes, the program now will create a config.txt
file in the local directory if it cannot find one (it will be
initialized with default settings) and it will save keybinding changes
to config.txt when the window is closed.